refactor(common): promote 'schnorr' primitives to the crypto_helper module
